moved ListItemProgressView from ListItemView, since it has more then 1 dependecy.  Also used in TransactionsScene separetly
