+++ Packages/Components/Sources/Lists/ListItemView.swift
@@ -1,3 +1,5 @@
+// Copyright (c). Gem Wallet. All rights reserved.
+
 import SwiftUI
 import Style
 
@@ -388,20 +390,3 @@ public enum TitleTagType {
     case image(Image)
 }
 
-public struct ListItemProgressView: View {
-
-    public let size: ControlSize
-    public let tint: Color
-
-    public init(size: ControlSize, tint: Color) {
-        self.size = size
-        self.tint = tint
-    }
-
-    public var body: some View {
-        ProgressView()
-            .controlSize(size)
-            .progressViewStyle(.circular)
-            .tint(tint)
-    }
-}
